+++ app/src/androidTest/java/com/pgtest/example/kii/fb_integration/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.pgtest.example.kii.fb3_integration;
+package com.pgtest.example.kii.fb_integration;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
