+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.pgtest.example.kii.fb3_integration" >
+    package="com.pgtest.example.kii.fb_integration" >
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
@@ -10,7 +10,7 @@
         android:theme="@style/AppTheme" >
         <meta-data android:name="com.facebook.sdk.ApplicationId" android:value="@string/facebook_app_id" />
         <activity
-            android:name="com.pgtest.example.kii.fb3_integration.MainActivity"
+            android:name="com.pgtest.example.kii.fb_integration.MainActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
